backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
gunicorn
deepface==0.0.79
opencv-python
numpy
mediapipe
twilio
python-dotenv